requirements.txt
aiohttp>=3.8.0
pyyaml>=6.0
requests>=2.28.0
click>=8.0.0
colorama>=0.4.6
jinja2>=3.1.0
python-dateutil>=2.8.0
